fix:fix mul pin

+++ dag/pool/datapin/pin_model.go
@@ -5,11 +5,12 @@ import (
 	"fmt"
 	"github.com/filedag-project/filedag-storage/dag/pool/datapin/types"
 	"github.com/filedag-project/filedag-storage/http/objectstore/uleveldb"
+	"golang.org/x/xerrors"
 	"sync"
 )
 
 type BlockPin struct {
-	mu sync.RWMutex
+	mu sync.Mutex
 	DB *uleveldb.ULevelDB
 }
 
@@ -17,9 +18,12 @@ const dagPoolPin = "dagPoolPin/"
 
 // AddPin add pin
 func (p *BlockPin) AddPin(pin types.Pin) error {
-	defer p.mu.RUnlock()
+	p.mu.Lock()
+	defer p.mu.Unlock()
+	if p.Has(pin.Cid.String()) {
+		return nil
+	}
 	cidCode := sha256String(pin.Cid.String())
-	p.mu.RLock()
 	err := p.DB.Put(dagPoolPin+cidCode, pin)
 	if err != nil {
 		return err
@@ -29,6 +33,11 @@ func (p *BlockPin) AddPin(pin types.Pin) error {
 
 // RemovePin remove pin
 func (p *BlockPin) RemovePin(cid string) error {
+	p.mu.Lock()
+	defer p.mu.Unlock()
+	if p.Has(cid) {
+		return xerrors.Errorf("not found : %v", cid)
+	}
 	cidCode := sha256String(cid)
 	err := p.DB.Delete(dagPoolPin + cidCode)
 	if err != nil {
@@ -38,6 +47,8 @@ func (p *BlockPin) RemovePin(cid string) error {
 }
 
 func (p *BlockPin) QueryPin(cid string) (*types.Pin, error) {
+	p.mu.Lock()
+	defer p.mu.Unlock()
 	cidCode := sha256String(cid)
 	pin := &types.Pin{}
 	err := p.DB.Get(dagPoolPin+cidCode, &pin)
@@ -45,6 +56,15 @@ func (p *BlockPin) QueryPin(cid string) (*types.Pin, error) {
 		return nil, err
 	}
 	return pin, nil
+}
+func (p *BlockPin) Has(cid string) bool {
+	cidCode := sha256String(cid)
+	pin := &types.Pin{}
+	err := p.DB.Get(dagPoolPin+cidCode, &pin)
+	if err != nil {
+		return false
+	}
+	return true
 }
 
 func NewBlockPin(db *uleveldb.ULevelDB) (BlockPin, error) {
